feat: add backend code about edit farm and delete farm

+++ backend/src/main/java/com/nova/backend/preset/repository/PresetRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface PresetRepository extends JpaRepository<PresetEntity,Integer> {
+public interface PresetRepository extends JpaRepository<PresetEntity,Long> {
     // UserId가 userId or NULL 인 경우
     @Query("SELECT p FROM PresetEntity p WHERE p.user.userId = :userId OR p.user IS NULL")
     List<PresetEntity> findCustomOrSystemPresets(@Param("userId") Long userId);
